updated auth mutation
